requirements.txt

streamlit==1.28.0
openai==1.3.0
langchain==0.0.335
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
reportlab==3.6.12
markdown==3.4.1
pyperclip